requirements.txt
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
bandit==1.9.2
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.18
decorator==5.2.1
defusedxml==0.8.0rc2
Django==6.0
django-allauth==65.13.1
django-axes==8.0.0
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-defender==0.9.8
django-environ==0.12.0
django-oauth-toolkit==3.1.0
django-ratelimit==4.1.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework-jsonapi==8.1.0
djangorestframework-stubs==3.16.6
djangorestframework_simplejwt==5.5.1
executing==2.2.1
fastjsonschema==2.21.2
fqdn==1.5.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==1.0.dev3
idna==3.11
inflection==0.5.1
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
jwcrypto==1.5.6
lark==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
mysql-connector-python==9.5.0
mysqlclient==2.2.7
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.0
notebook_shim==0.2.4
oauthlib==3.3.1
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
platformdirs==4.5.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.3
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pypandoc==1.16.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.4
stack-data==0.6.3
stevedore==5.6.0
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.3
traitlets==5.14.3
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
whitenoise==6.11.0
widgetsnbextension==4.0.15
yt-dlp==2025.12.13.232949.dev0
